Improve error handling for WarmUp() (#28)
Users have reported an issue where importing the module on PowerShell's
startup when offline may result in an unhandled exception. The call
stack shows that the issue is from the async `Find-WinGetPackage` call
in `WarmUp()`. This is bizarre given that it's wrapped in a try-catch
block!

> Async void methods have different error-handling semantics. When an
exception is thrown out of an async Task or async Task<T> method, that
exception is captured and placed on the Task object. With async void
methods, there is no Task object, so any exceptions thrown out of an
async void method will be raised directly on the SynchronizationContext
that was active when the async void method started.

Reference:
https://learn.microsoft.com/en-us/archive/msdn-magazine/2013/march/async-await-best-practices-in-asynchronous-programming

Given that information, the fix here is to change the method signature
to return a `Task` rather than being `void`. We're then able to attach a
`ContinueWith()` block to the returned task to detect if the task threw
an exception (`Task.IsFaulted`). In this scenario, we set `_warmedUp` to
false.

Rather than trying to warm up again in the future, `FindPackages()` now
exits early, preventing future calls to `Find-WinGetPackage`. This
essentially disables the package.

Closes #20