Search only by command and update main error handling (#29)
Makes the WinGet search more strict in that it only searches for a
strict match by command. This removes the blocking calls from various
searches, which was the root cause of #24.

This also prevents the scenario where a package is suggested, but the
newly installed package doesn't actually fix the issue of a command not
being found. In searching only by command, the newly installed package
is guaranteed to introduce the missing command unless the WinGet package
catalog was wrong.

Additionally, the `catch` block was updated to remove the known MSIX
issue, which has now been fixed. Instead, we display the caught
exception. This should help further debug issues like #23.

Closes #13
Closes #24